--- a/spec/ameba/ast/traverse_spec.cr
+++ b/spec/ameba/ast/traverse_spec.cr
@@ -8,7 +8,7 @@ module Ameba::AST
     {% for name in NODE_VISITORS %}
       describe "{{name}}" do
         it "allow to visit {{name}} node" do
-          visitor = {{name}}Visitor.new rule, source
+          visitor = Visitor.new rule, source
           nodes = Crystal::Parser.new("").parse
           nodes.accept visitor
         end