--- a/src/ameba/rules/method_names.cr
+++ b/src/ameba/rules/method_names.cr
@@ -32,7 +32,7 @@ module Ameba::Rules
   # ```
   struct MethodNames < Rule
     def test(source)
-      AST::DefVisitor.new self, source
+      AST::Visitor.new self, source
     end
 
     def test(source, node : Crystal::Def)