--- a/src/ameba/rules/negated_conditions_in_unless.cr
+++ b/src/ameba/rules/negated_conditions_in_unless.cr
@@ -22,7 +22,7 @@ module Ameba::Rules
   #
   struct NegatedConditionsInUnless < Rule
     def test(source)
-      AST::UnlessVisitor.new self, source
+      AST::Visitor.new self, source
     end
 
     def test(source, node : Crystal::Unless)