--- a/src/ameba/rules/predicate_name.cr
+++ b/src/ameba/rules/predicate_name.cr
@@ -24,7 +24,7 @@ module Ameba::Rules
   #
   struct PredicateName < Rule
     def test(source)
-      AST::DefVisitor.new self, source
+      AST::Visitor.new self, source
     end
 
     def test(source, node : Crystal::Def)