--- a/src/ameba/rules/unless_else.cr
+++ b/src/ameba/rules/unless_else.cr
@@ -38,7 +38,7 @@ module Ameba::Rules
   #
   struct UnlessElse < Rule
     def test(source)
-      AST::UnlessVisitor.new self, source
+      AST::Visitor.new self, source
     end
 
     def test(source, node : Crystal::Unless)