--- a/src/ameba/source.cr
+++ b/src/ameba/source.cr
@@ -15,6 +15,7 @@ module Ameba
     getter errors = [] of Error
     getter path : String?
     getter content : String
+    getter ast : Crystal::ASTNode?
 
     def initialize(@content : String, @path = nil)
     end
@@ -29,6 +30,13 @@ module Ameba
 
     def lines
       @lines ||= @content.split("\n")
+    end
+
+    def ast
+      @ast ||=
+        Crystal::Parser.new(content)
+                       .tap { |parser| parser.filename = @path }
+                       .parse
     end
   end
 end